move keyhelpers to helpers directory, name it the name of the export
